+++ kite-core/src/main/kotlin/com/tang/kite/datasource/DataSourceContext.kt
@@ -1,7 +1,5 @@
 package com.tang.kite.datasource
 
-import java.util.function.Supplier
-
 /**
  * @author Tang
  */
@@ -35,42 +33,4 @@ object DataSourceContext {
         context.remove()
     }
 
-    fun <T> with(key: String, block: () -> T): T {
-        try {
-            push(key)
-            return block()
-        } finally {
-            pop()
-        }
-    }
-
-    @JvmStatic
-    fun with(key: String, runnable: Runnable) {
-        try {
-            push(key)
-            runnable.run()
-        } finally {
-            pop()
-        }
-    }
-
-    @JvmStatic
-    fun <T> with(key: String, supplier: Supplier<T>): T {
-        try {
-            push(key)
-            return supplier.get()
-        } finally {
-            pop()
-        }
-    }
-
-}
-
-inline fun <T> withDataSource(key: String, block: () -> T): T {
-    try {
-        DataSourceContext.push(key)
-        return block()
-    } finally {
-        DataSourceContext.pop()
-    }
 }